+++ test/bagofcli.js
@@ -233,6 +233,52 @@ buster.testCase('cli - _postCommand', {
       commandsConfig = { somecommand: { args: [{ name: 'arg1', rules: [ 'number' ]}, { name: 'arg2', rules: [ 'number' ] }] } },
       result = bag._postCommand(args, commandsConfig);
     assert.equals(result, undefined);
+  },
+  'should log error message when command has invalid command option': function (done) {
+    this.mockConsole.expects('error').once().withExactArgs('Invalid options: <-s, --some-arg <someArg>> must be number'.red);
+    this.mockProcess.expects('exit').once().withExactArgs(1);
+    var args = ['123', { _name: 'somecommand', someArg: 'abcdef', parent: { _name: 'someparentcommand' } }],
+      commandsConfig = { somecommand: { options: [{
+        arg: '-s, --some-arg <someArg>',
+        rules: [ 'number' ]
+      }]}},
+      err, result;
+    try {
+      result = bag._postCommand(args, commandsConfig);
+    } catch (e) {
+      err = e;
+    }
+    assert.equals(result, undefined);
+    done(err);
+  },
+  'should return without error when there is no invalid command option': function (done) {
+    var args = ['123', { _name: 'somecommand', someArg: '12345', parent: { _name: 'someparentcommand' } }],
+      commandsConfig = { somecommand: { options: [{
+        arg: '-s, --some-arg <someArg>',
+        rules: [ 'number' ]
+      }]}},
+      err, result;
+    try {
+      result = bag._postCommand(args, commandsConfig);
+    } catch (e) {
+      err = e;
+    }
+    assert.equals(result, undefined);
+    done(err);
+  },
+  'should return without error when command option does not have any validation rule': function (done) {
+    var args = ['123', { _name: 'somecommand', someArg: '12345', parent: { _name: 'someparentcommand' } }],
+      commandsConfig = { somecommand: { options: [{
+        arg: '-s, --some-arg <someArg>'
+      }]}},
+      err, result;
+    try {
+      result = bag._postCommand(args, commandsConfig);
+    } catch (e) {
+      err = e;
+    }
+    assert.equals(result, undefined);
+    done(err);
   }
 });